Start of burst test using a subprocess.
Because GRC logs errors directly to the stderr and python does not allow direct reading of the stderr, the body of the start of burst test was placed in [its own file](/python/qa_crimson_burst_dummy.py). It is then called using a subprocess in the [loopback test](/python/qa_crimson_loopback.py) whose stderr is captured and asserted for the appropriate error message.

In order for this to work, the followng must be done prior to executing `make test` in the build sequence:
- Copy [/qa_crimson_burst_dummy.sh] to [/build/python/].
- Replace the generated [/build/python/qa_crimson_loopback_test.sh] with the (version in the root)[/qa_crimson_loopback.sh].

This is very messy; looking for a better solution.